Refactor staff views into their own blueprint. Included delete users in staff list - in process for TN-138
