+++ portal/factories/app.py
@@ -36,6 +36,7 @@ from ..views.procedure import procedure_api
 from ..views.portal import portal
 from ..views.organization import org_api
 from ..views.scheduled_job import scheduled_job_api
+from ..views.staff import staff
 from ..views.tou import tou_api
 from ..views.truenth import truenth_api
 from ..views.user import user_api
@@ -59,6 +60,7 @@ DEFAULT_BLUEPRINTS = (
     procedure_api,
     portal,
     scheduled_job_api,
+    staff,
     truenth_api,
     tou_api,
     user_api,)
